Pass favorites as props to GlobalProvider in App.js
